requirements.txt
Flask==2.0.2
Pillow==8.4.0
stepic==0.5.0
cryptography==3.4.8
Werkzeug==2.0.2
gunicorn==20.1.0  # Adding gunicorn for production server
